pyproject.toml
[tool.poetry]
name = "python-example-jwt"
version = "0.1.0"
description = "JWT Auth example"
authors = ["David Rosero Calle <davidroserocalle@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
fastapi = "^0.110.0"
pyjwt = "^2.8.0"
python-decouple = "^3.8"
pydantic = {extras = ["email"], version = "^2.6.3"}
gunicorn = "^21.2.0"
pytest = "^8.0.2"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
